+++ examples/wip/viewport-animation-first-person-map/app.js
@@ -12,7 +12,7 @@ import {
   WebMercatorViewport,
   PolygonLayer,
   PointCloudLayer,
-  ViewportLinearInterpolator
+  LinearInterpolator
 } from 'deck.gl';
 
 import TripsLayer from '../../trips/trips-layer';
@@ -42,8 +42,6 @@ const LIGHT_SETTINGS = {
   numberOfLights: 2
 };
 
-const TRANSITION_PROPS = ['longitude', 'latitude', 'zoom', 'bearing', 'pitch', 'position'];
-
 const DEFAULT_VIEWPORT_PROPS = {
   longitude: -74,
   latitude: 40.72,
@@ -57,6 +55,10 @@ const DEFAULT_VIEWPORT_PROPS = {
   // direction: [-0.9, 0.5, 0], // Which direction is camera looking at, default origin
   up: [0, 0, 1] // Defines up direction, default positive y axis
 };
+
+const transitionInterpolator = new LinearInterpolator(
+  ['longitude', 'latitude', 'zoom', 'bearing', 'pitch', 'position']
+);
 
 class Root extends Component {
   constructor(props) {
@@ -316,7 +318,7 @@ class Root extends Component {
           height={viewportProps.height}
           onViewportChange={this._onViewportChange}
           transitionDuration={transitionDuration}
-          transitionInterpolator={new ViewportLinearInterpolator(TRANSITION_PROPS)}
+          transitionInterpolator={transitionInterpolator}
         >
           <DeckGL
             id="first-person"
