+++ src/core/lib/transition/index.js
@@ -1,0 +1,5 @@
+export {default as TransitionInterpolator} from './transition-interpolator';
+export {default as ViewportFlyToInterpolator} from './viewport-fly-to-interpolator';
+export {default as LinearInterpolator} from './linear-interpolator';
+
+export {extractViewportFrom} from './transition-utils';
