+++ src/core/lib/transition/transition-utils.js
@@ -1,0 +1,43 @@
+import {mod} from '../../math/utils';
+
+const VIEWPORT_PROPS =
+  ['width', 'height', 'longitude', 'latitude', 'zoom', 'bearing', 'pitch', 'position'];
+
+const WRAPPED_ANGULAR_PROPS = {
+  longitude: 1,
+  bearing: 1
+};
+
+export function lerp(start, end, step) {
+  if (Array.isArray(start)) {
+    return start.map((element, index) => {
+      return lerp(element, end[index], step);
+    });
+  }
+  return step * end + (1 - step) * start;
+}
+
+export function isValid(prop) {
+  return Number.isFinite(prop) || Array.isArray(prop);
+}
+
+export function isWrappedAngularProp(propName) {
+  return WRAPPED_ANGULAR_PROPS[propName];
+}
+
+// TODO/xiaoji: This should be merged with the controller's prop constraint system
+export function extractViewportFrom(props) {
+  const viewport = {};
+  VIEWPORT_PROPS.forEach((key) => {
+    const value = props[key];
+    if (isValid(value)) {
+      viewport[key] = value;
+      // Normalize longitude and bearing into [-180, 180) range
+      // This gurantees the props are in same range when they are interpolated.
+      if (isWrappedAngularProp(key)) {
+        viewport[key] = mod(value + 180, 360) - 180;
+      }
+    }
+  });
+  return viewport;
+}
