+++ src/index.js
@@ -56,7 +56,7 @@ export {
 // Transition bindings
 export {
   TRANSITION_EVENTS,
-  ViewportLinearInterpolator,
+  LinearInterpolator,
   ViewportFlyToInterpolator
   // TransitionManager (Does the TransitionManager need to be exported?)
 } from './core';
